CPP02

+++ CPP02/ex01/main.cpp
@@ -1,17 +1,27 @@
-#include "Fixed.hpp"
-#include <climits>
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   main.cpp                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aahlyel <aahlyel@student.1337.ma>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/10/23 09:36:24 by aahlyel           #+#    #+#             */
+/*   Updated: 2023/10/23 09:36:25 by aahlyel          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-int main( void )
-{
+#include "Fixed.hpp"
+
+int main( void ) {
 	Fixed a;
 	Fixed const b( 10 );
 	Fixed const c( 42.42f );
 	Fixed const d( b );
 	a = Fixed( 1234.4321f );
-	std::cout << "a is " << a.toFloat() << std::endl;
-	std::cout << "b is " << b.toFloat() << std::endl;
-	std::cout << "c is " << c.toFloat() << std::endl;
-	std::cout << "d is " << d.toFloat() << std::endl;
+	std::cout << "a is " << a << std::endl;
+	std::cout << "b is " << b << std::endl;
+	std::cout << "c is " << c << std::endl;
+	std::cout << "d is " << d << std::endl;
 	std::cout << "a is " << a.toInt() << " as integer" << std::endl;
 	std::cout << "b is " << b.toInt() << " as integer" << std::endl;
 	std::cout << "c is " << c.toInt() << " as integer" << std::endl;
